Send private Telegram messages to users
Use the user's phone number attribute to send a private Telegram message
for notification.
To set up, the notificator still requires the chatId attribute - it will
be used to retrieve the user_id by using the user's phone number.
Users will have to activate the bot - this is done in the same manner as
any other Telegram bot.