Telegram: cache chat id per phone number
Retriving the chat id per each notification is inefficiant.
This patch adds a simple cache, keyed by phone number and storing the
chat id.
Upon retrieval, the chat id is stored in the cache and will be used for
the time period that is set in the config with:
notificator.telegram.cacheRetention

Cache entries which expired aren't cleaned up - they'll stay around
until the server is restarted or the user is paged again and the cache
will be refreshed.